+++ app/Imports/Srd/SrdRoutesImport.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Imports;
+namespace App\Imports\Srd;
 
 use App\Models\Srd\SrdRoute;
 use Illuminate\Support\Collection;
